Decouple AbstractController for test
