+++ src/AppBundle/Tests/Controller/SwaggerValidator/Parameters/HeaderParameter/StringHeaderParameterValidator.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace AppBundle\Tests\Controller\SwaggerValidator\Parameters\HeaderParameter;
 
 use AppBundle\Tests\Controller\SwaggerValidator\Parameters\AbstractParameterValidator;
